requirements.txt
# Core dependencies
PyYAML>=6.0
click>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional dependencies for enhanced functionality
colorama>=0.4.0  # Colored output
rich>=12.0.0     # Rich console output
schedule>=1.2.0  # Task scheduling